Drop retail_price from pricing policy interface
Since pricing policies were implemented, the retail_price attribute was
a part of the interface, but not used. As the retail price is only ever
used for display purposes, it doesn't semantically belong into the
pricing policy, which is more concerned with business logic.

The retail price is still easily accessible as PurchaseInfo pass the
stockrecord, where it can be accessed in e.g. templates.