Differentiate between no tax and zero tax
There is a distinction between no tax (e.g. in B2B applications) and a
tax of 0% (e.g. tax-exempt products). The latter still allows access to
the incl_tax and tax attributes of the Price instances.

This commit changes the default strategy from no tax to zero tax,
updates the tests accordingly, and clarifies the
documentation for the NoTax mixin.

Note that accessing any of the backwards compatible fields on the
Unavailable pricing policy will now raise the TaxNotKnown exception.