Price class: Always return quantized decimals
Quantization should happen as late as possible in calculations, and
hence should be the price class' concern. It not always returns
quantized decimals.

The exponent should be made configurable eventually, but right not other
parts of the codebase perform quantization and tax-calculation(!), which
needs to be addressed first. Grepping for 'quantize' shows what I mean.